Change methodology for running unit tests to account for binary wheels.
