+++ .github/workflows/build.yaml
@@ -27,6 +27,10 @@ jobs:
         with:
           name: sdist
           path: 'dist/**'
+      - uses: actions/upload-artifact@v3
+        with:
+          name: test-files
+          path: 'spotfire/test/files/**'
       - name: Dynamic Elements
         id: dynamic
         run: |
@@ -59,10 +63,15 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python .github/scripts/parse_pyproject_toml.py --section build-system --key requires --array --string > pyproj-requires.txt
-          pip install -r pyproj-requires.txt
+          pip install auditwheel -r pyproj-requires.txt
       - name: Build Wheel
         run: |
           python setup.py bdist_wheel
+      - name: Convert Wheel to Manylinux
+        if: ${{ matrix.operating-system == 'ubuntu-latest' }}
+        run: |
+          mv dist/*.whl .
+          auditwheel repair -w dist ./*.whl
       - uses: actions/upload-artifact@v3
         with:
           name: wheel-${{ matrix.python-version }}-${{ matrix.operating-system }}
@@ -70,20 +79,22 @@ jobs:
   test:
     name: Test
     needs: [lint, build]
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.operating-system }}
     strategy:
       matrix:
         python-version: ${{ fromJson(needs.lint.outputs.python-versions) }}
+        operating-system: ['ubuntu-latest', 'windows-latest']
         test-environment: ${{ fromJson(needs.lint.outputs.test-environments) }}
       fail-fast: false
     steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: recursive
       - uses: actions/download-artifact@v3
         with:
-          name: wheel-${{ matrix.python-version }}-ubuntu-latest
+          name: wheel-${{ matrix.python-version }}-${{ matrix.operating-system }}
           path: dist
+      - uses: actions/download-artifact@v3
+        with:
+          name: test-files
+          path: test-files
       - name: Set Up Python
         uses: actions/setup-python@v4
         with:
@@ -96,10 +107,10 @@ jobs:
         run: |
           python -m spotfire.test
         env:
-          RESULTS_NAME: ${{ matrix.python-version }}-${{ matrix.test-environment }}
+          RESULTS_NAME: ${{ matrix.python-version }}-${{ matrix.operating-system }}-${{ matrix.test-environment }}
+          TEST_FILES_DIR: ${{ env.GITHUB_WORKSPACE }}/test-files
       - uses: actions/upload-artifact@v3
         if: ${{ always() }}
         with:
-          name: test-results-${{ matrix.python-version }}-${{ matrix.test-environment }}
+          name: test-results-${{ matrix.python-version }}-${{ matrix.operating-system }}-${{ matrix.test-environment }}
           path: 'results-*.xml'
-
