+++ pylintrc
@@ -3,7 +3,7 @@
 # A comma-separated list of package or module names from where C extensions may
 # be loaded. Extensions are loading into the active Python interpreter and may
 # run arbitrary code
-extension-pkg-whitelist=
+extension-pkg-allow-list=spotfire.sbdf,spotfire.codesign
 
 # Add files or directories to the blacklist. They should be base names, not
 # paths.
