+++ spotfire/test/test_sbdf.py
@@ -2,7 +2,6 @@
 
 import datetime
 import decimal
-import os
 import unittest
 import tempfile
 import pkg_resources
@@ -12,6 +11,7 @@ import geopandas
 
 import spotfire
 from spotfire import sbdf
+from spotfire.test import utils
 
 
 class SbdfTest(unittest.TestCase):
@@ -20,7 +20,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_0(self):
         """Reading simple SBDF files should work."""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/0.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/0.sbdf"))
         self.assertEqual(dataframe.shape, (0, 12))
 
         def verify(dict_, pre, post):
@@ -56,7 +56,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_1(self):
         """Reading simple SBDF files should work."""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/1.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/1.sbdf"))
         self.assertEqual(dataframe.shape, (1, 12))
         self.assertEqual(dataframe.at[0, "Boolean"], False)
         self.assertEqual(dataframe.at[0, "Integer"], 69)
@@ -72,7 +72,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_100(self):
         """Reading simple SBDF files should work."""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/100.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/100.sbdf"))
         self.assertEqual(dataframe.shape, (100, 12))
         self.assertEqual(dataframe.get("Boolean")[0:6].tolist(), [False, True, None, False, True, None])
         self.assertEqual(dataframe.get("Integer")[0:6].dropna().tolist(), [69.0, 73.0, 75.0, 79.0])
@@ -90,7 +90,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_10001(self):
         """Reading simple SBDF files should work."""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/10001.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/10001.sbdf"))
         self.assertEqual(dataframe.shape, (10001, 12))
         # Check the values in the first row
         self.assertEqual(dataframe.at[0, "Boolean"], False)
@@ -119,7 +119,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_write_geodata(self):
         """Test that geo-encoded data is properly converted to/from GeoDataFrame"""
-        gdf = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/NACountries.sbdf")
+        gdf = sbdf.import_data(utils.get_test_data_file("sbdf/NACountries.sbdf"))
         self.assertIsInstance(gdf, pandas.DataFrame)
         self.assertIsInstance(gdf, geopandas.GeoDataFrame)
 
@@ -142,7 +142,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_write_unicode(self):
         """Test that unicode string arrays are properly written"""
-        udf = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/unicode.sbdf")
+        udf = sbdf.import_data(utils.get_test_data_file("sbdf/unicode.sbdf"))
         with tempfile.TemporaryDirectory() as tempdir:
             sbdf.export_data(udf, f"{tempdir}/test.sbdf")
             udf2 = sbdf.import_data(f"{tempdir}/test.sbdf")
@@ -151,7 +151,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_read_write_alltypes(self):
         """Test that all data types can be properly roundtripped read/write"""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/alltypes.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/alltypes.sbdf"))
         with tempfile.TemporaryDirectory() as tempdir:
             sbdf.export_data(dataframe, f"{tempdir}/test.sbdf")
             df2 = sbdf.import_data(f"{tempdir}/test.sbdf")
@@ -177,7 +177,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_get_spotfire_types(self):
         """All types should be reported properly"""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/alltypes.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/alltypes.sbdf"))
         type_names = spotfire.get_spotfire_types(dataframe)
         self.assertEqual(type_names["ColumnBoolean"], "Boolean")
         self.assertEqual(type_names["ColumnDate"], "Date")
@@ -194,7 +194,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_set_spotfire_types(self):
         """Setting SBDF types should work properly"""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/alltypes.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/alltypes.sbdf"))
         # set a single column by name
         spotfire.set_spotfire_types(dataframe, {"ColumnLong": "Integer"})
         self.assertEqual(spotfire.get_spotfire_types(dataframe)["ColumnLong"], "Integer")
@@ -238,7 +238,7 @@ class SbdfTest(unittest.TestCase):
 
     def test_import_export_alltypes(self):
         """Verify all types properly export and re-import with the proper Spotfire type"""
-        dataframe = sbdf.import_data(f"{os.path.dirname(__file__)}/files/sbdf/alltypes.sbdf")
+        dataframe = sbdf.import_data(utils.get_test_data_file("sbdf/alltypes.sbdf"))
         with tempfile.TemporaryDirectory() as tempdir:
             sbdf.export_data(dataframe, f"{tempdir}/output.sbdf")
             new_df = sbdf.import_data(f"{tempdir}/output.sbdf")
@@ -364,8 +364,9 @@ class SbdfTest(unittest.TestCase):
             self.assertEqual(new_df_types["x"], default_type)
         # if default is None expect failure
         else:
-            with self.assertRaises(sbdf.SBDFError):
-                sbdf.export_data(dataframe, "output.sbdf")
+            with tempfile.TemporaryDirectory() as tempdir:
+                with self.assertRaises(sbdf.SBDFError):
+                    sbdf.export_data(dataframe, f"{tempdir}/output.sbdf")
 
         # validate expected passing cases
         for df_type in pass_types:
