+++ spotfire/test/utils.py
@@ -1,6 +1,8 @@
 """Utility classes for integrating unit test results into Jenkins."""
 
 
+import inspect
+import os
 import sys
 import xml.dom.minidom
 import xmlrunner
@@ -38,3 +40,19 @@ class PythonVersionModifiedResult(xmlrunner.result._XMLTestResult):  # pylint: d
 
 
 PYTHON_VERSION = f"py{sys.version_info.major}{sys.version_info.minor}"
+
+
+def get_test_data_file(name):
+    """Determine the filename of a test data file.  It is possible for the data files to be separated from the
+    ``spotfire.test.files`` package, especially for testing binary wheels in the CI environment; to look in an
+    alternate location, set the ``TEST_FILES_DIR`` environment variable.
+
+    :param name: the basename of the test file
+    :return: the full filename of the test file
+    """
+    test_dir, test_file = os.path.split(inspect.stack()[1].filename)
+    if "TEST_FILES_DIR" in os.environ:
+        files_dir = os.getenv("TEST_FILES_DIR")
+    else:
+        files_dir = os.path.join(test_dir, "files")
+    return os.path.join(files_dir, name)
